requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
stripe>=8.0.0
psycopg2-binary>=2.9.0